chore: fix pact test for system account
The system account creation uses the Location header to extract the user's UUID. This was missing in the pact test, but was not apparent because the assertion checked a value that was guaranteed to be correct (a value that was not extracted from the HTTP request). The only value that is dynamic from the HTTP test was the UUID, which was not covered by the test.

This change addresses the gap.